+++ packages/pluggableWidgets/tooltip-web/src/utils/useDocumentKeyDown.ts
@@ -1,6 +1,6 @@
 import { useEffect } from "react";
 
-export default function useDocumentKeyDown(callback: (event: KeyboardEvent) => void): void {
+export function useDocumentKeyDown(callback: (event: KeyboardEvent) => void): void {
     useEffect(() => {
         document.addEventListener("keydown", callback);
         return () => {