+++ packages/pluggableWidgets/datagrid-web/CHANGELOG.md
@@ -6,9 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+## [2.8.0] - 2023-08-10
+
 ### Changed
 
 -   We improved accessibility of the Datagrid widget.
+
 -   We changed the DOM Structure for filter options to appear inline with the container in order to make the component more accessible.
 
 ### Fixed